Make "context" a "common" field name instead of being retricted to ILoggingEvent
Issue #658
